requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
easyocr>=1.7.0
python-docx>=1.1.0
Pillow>=10.0.0
imagehash>=4.3.1
numpy>=1.24.0
